Enable SSR  (#2)
Moves the initial EditorView creation into the render function of the ProseMirror component. In order to do this, we have to construct the EditorView without a mount initially, which ProseMirror handles by creating an empty div. To work around the EditorView constructor accessing window and document, we introduce some super dumb stubs for window and document that we set before calling super and then clean up (if we're in SSR).